Text wrapping - at long last
At long last, I have come to understand how to take a single damned string and, given a maximum length for a row of text, split it into multiple rows that take into consideration most any whitespace configuration. There is also included the function that led to this understanding, as a reference point. I know I will never get that clean with what I have, but I wanted to develop an understanding for how that function did its job.